backend/requirements.txt
Flask==3.1.1
SQLAlchemy==2.0.41
authlib==1.5.2
apscheduler==3.11.0
requests==2.32.3
python-dotenv==1.1.0
